requirements.txt
torch
transformers
datasets
pandas
scikit-learn
python-dotenv
requests
motor
dnspython
